lexer: Don't assume double if more than one dot in string
This allows matches with more than one dot in the string, such as
"org.gnome.Calculator". Previously, we were matching double if there
were any dots in the string. This allows to match as a string type
if there are more than one dot while preserving double for just one dot.